Infinite list implemented. BUG: Empty list when rotating or back from detail. TODO: Manage Network Error and Retry.

+++ app/src/main/java/com/alfonsocastro/superhero/ui/herodetail/HeroDetailFragment.kt
@@ -14,7 +14,7 @@ import com.alfonsocastro.superhero.databinding.FragmentHeroDetailBinding
 class HeroDetailFragment : Fragment() {
 
     private val viewModel: HeroDetailViewModel by viewModels {
-        HeroDetailViewModel.HeroDetailViewModelFactory(args.heroId)
+        HeroDetailViewModel.HeroDetailViewModelFactory(args.heroId.toInt())
     }
 
     private val args: HeroDetailFragmentArgs by navArgs()
